Drop texture. Use garden colors in polyptych borders and link underlines.
